+++ examples/resources/iosxe_save_config/resource.tf
@@ -1,0 +1,2 @@
+resource "iosxe_save_config" "example" {
+}