+++ internal/provider/provider.go
@@ -275,6 +275,7 @@ func (p *IosxeProvider) Configure(ctx context.Context, req provider.ConfigureReq
 func (p *IosxeProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewRestconfResource,
+		NewSaveConfigResource,
 		NewAAAResource,
 		NewAAAAccountingResource,
 		NewAAAAuthenticationResource,